+++ build.rs
@@ -1,0 +1,6 @@
+//! Custom hooks and triggers for compiling buildbtw
+
+fn main() {
+    // Re-check GraphQL queries if the GraphQL schema changes.
+    println!("cargo::rerun-if-changed=src/gitlab/graphql_schema.json");
+}